Fix: Swoll hides meat in bag
Having the Swoll in your bag causes the first meat in your inventory to
be skipped and not printed, because the variable tracking which line to
print on and the variable used to iterate through the hero's inventory
were the same.  Split them into two variables so iterating through the
bag will start at items[0] even if the Swoll line has been printed.